+++ crates/usvg/src/writer.rs
@@ -628,14 +628,6 @@ fn conv_element(node: &Node, is_clip_path: bool, opt: &XmlOptions, xml: &mut Xml
                     .map(|filter| format!("url(#{}{})", prefix, filter.id))
                     .collect();
                 xml.write_svg_attribute(AId::Filter, &ids.join(" "));
-
-                if let Some(ref fill) = g.filter_fill_paint() {
-                    write_paint(AId::Fill, fill, opt, xml);
-                }
-
-                if let Some(ref stroke) = g.filter_stroke_paint() {
-                    write_paint(AId::Stroke, stroke, opt, xml);
-                }
             }
 
             if g.opacity != Opacity::ONE {
@@ -678,7 +670,7 @@ fn conv_element(node: &Node, is_clip_path: bool, opt: &XmlOptions, xml: &mut Xml
             xml.end_element();
         }
         NodeKind::Text(_) => {
-            log::warn!("text must be converted into paths");
+            log::warn!("Text must be converted into paths.");
         }
     }
 }
@@ -854,10 +846,6 @@ impl XmlWriterExt for XmlWriter {
             match input {
                 filter::Input::SourceGraphic => "SourceGraphic",
                 filter::Input::SourceAlpha => "SourceAlpha",
-                filter::Input::BackgroundImage => "BackgroundImage",
-                filter::Input::BackgroundAlpha => "BackgroundAlpha",
-                filter::Input::FillPaint => "FillPaint",
-                filter::Input::StrokePaint => "StrokePaint",
                 filter::Input::Reference(ref s) => s,
             },
         );